backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
pillow==10.0.1
python-multipart==0.0.6
cors==1.0.1
pydantic==2.4.2
scikit-learn==1.3.0
matplotlib==3.7.2